0f38d82 Merge topic 'axis-fixes' into master
6024ed5 Deprecate SetTickPositions/SetTickLabels for new API
603e222 Perform a deep copy of custom axis positions/labels
b67dd64 Merge topic '0013430_change-exodus-block-names' into master
de2cc45 Merge topic 'amr-ghost-data-generation' into master
685b3d5 Merge topic 'clang-analyzer-fixes' into master
0172a13 COMP: fixed various dead writes found by clang static analyzer
f94602c COMP: fixed dead write (that was itself a warning workaround?)
4b78e72 BUG: fixed null deref by testing pointer first
154371d BUG: fixed null dereference found by clang static analyzer
9d43621 Remove the Size... prefix from Exodus block names
9693998 Merge topic 'clang-analyzer-fixes' into master
4f488c0 BUG: added default case to switch to fix invalid read
f6b39db COMP: fixed a few dead reads and dead writes
529056c Merge topic 'axis-fixes' into master
e25f4c1 Separate out custom tick labels, add more testing
...

+++ Common/DataModel/vtkAMRDataInternals.cxx
@@ -53,7 +53,7 @@ void vtkAMRDataInternals::PrintSelf(ostream& os, vtkIndent indent)
 void vtkAMRDataInternals::Insert(unsigned int index, vtkUniformGrid* grid)
 {
   this->Blocks.push_back(Block(index,grid));
-  int i = this->Blocks.size()-2;
+  int i = static_cast<int>(this->Blocks.size())-2;
   while( i>=0 && this->Blocks[i].Index > this->Blocks[i+1].Index)
     {
     std::swap(this->Blocks[i],this->Blocks[i+1]);
